Fix recommendations page so the empty recommendations picture is shown once all recommendations removed
